Protect the chat controller

+++ Models/AppSettings.cs
@@ -18,7 +18,7 @@
     public record JWT {
         public string Issuer { get; set; } = string.Empty;
         public string Audience { get; set; } = string.Empty;
-        public ushort ExpirationMinutes { get; set; }
+        public uint ExpirationMinutes { get; set; }
         public string SecretKey { get; set; } = string.Empty;
     }
 }
